requirements.txt
# YMYL Audit Tool - Simplified Version
# Minimal dependencies for core functionality

# Web Framework
streamlit>=1.28.0

# Web Scraping & Content Extraction
requests>=2.31.0
beautifulsoup4>=4.12.0

# AI Processing
openai>=1.0.0

# Document Export
python-docx>=0.8.11

# Utilities
pytz>=2023.3
